Be more paranoid about changes to disk
When conducting operations on the same localStorage from different
runtimes (ie, pages loaded in different tabs, or the background and
foreground pages of a browser extension) it's important to check for
changes to `this.records` before searching or saving that list.

Todo: let the store listen for 'storage' events and update the records
list automatically.